refactor: export IOption type from types
