+++ src/components/select/Select.tsx
@@ -8,9 +8,10 @@ import {
 } from 'react-select';
 
 import { type HelpTextProps } from 'src/components/help-text/HelpText';
+import { IOption } from 'src/types/IOption';
 import { Size } from 'src/types/Size';
 
-import { IOption, StyledReactSelect } from './_StyledReactSelect';
+import { StyledReactSelect } from './_StyledReactSelect';
 
 type RequiredProps = 'options' | 'value';
