+++ src/types/IOption.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export interface IOption<V extends string = string> {
+  icon?: ReactNode;
+  isDisabled?: boolean;
+  label: string;
+  labelDescription?: string;
+  value: V;
+}
